中秋快乐-itchat-输出更友好
The Midautumn Festival is the traditional festival in China.In this day,people ,espically family members will have a happy get-together.So Chinese people vaule this festival for its important meaning of "reunion".And the mooncake is the symbolic food.It represents the "reunion" just like the full moon.
Chinese Mid-Autumn Festival is one of the four major traditional festivals.中秋节是中国四大传统节日之一。
There are evening meal, people who work outside the home have to come back happy. After dinner, people lit lanterns, usually red lanterns round. The children will happily play their toy lanterns.
晚上有一顿美餐，离家在外工作的人也要回来团圆。晚饭后，人们点亮灯笼，一般是红色的圆灯笼。孩子们会高高兴兴地玩他们的玩具灯笼。
You Yuanyou the moon in the evening, people eating at the same time to celebrate the Mid-Autumn Festival special foods - cakes. People with the past, look to the future. It is said that there are dragon in the sky, it is necessary to swallow the moon. In order to protect the moon, the children have to come up with a large ring of the dragons away.
晚上月亮又圆又大，人们在赏月的同时吃着中秋节特别的食品——月饼。人们在一起回顾过去，展望未来。据说天上有一条龙，它要把月亮吞下去。为了保护月亮，孩子们要弄出很大的响动把龙吓跑。